파일구조 변경 + 랜딩페이지 변경
1. 컴포넌트 단위별로 파일 따로 보관
2. 랜딩페이지 제작 완료 : 현재는 로그인 회원가입 눌르면 대쉬보드로 이동되게 만들었음
3. 레이아웃 변경
- 컴포넌트/레이아웃/index.tsx에서 레이아웃 사용/사용X 상태 설정가능
4. 로그인 모달 만드는 중
