+++ Frontend/package.json
@@ -9,12 +9,13 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@svgr/webpack": "^8.1.0",
     "all": "^0.0.0",
     "chart.js": "^4.4.3",
     "docx-to-html": "^1.0.1",
     "draft-js": "^0.11.7",
     "eslint-plugin-jsx-a11y": "^6.8.0",
-    "next": "^14.2.3",
+    "next": "^14.2.5",
     "react": "^18.3.1",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.3.1",
