+++ Frontend/src/app/layout.tsx
@@ -1,29 +1,23 @@
 import type { Metadata } from 'next';
 import './globals.css';
-import Navbar from '@/components/Navbar';
-import Header from '@/components/Header';
+import Layout from '../components/layout/index';
 
 export const metadata: Metadata = {
   title: 'Create Next App',
   description: 'Generated by create next app',
 };
 
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
+
+export default function requestLayout({ children }: Readonly<{ children: React.ReactNode }>) {
   return (
     <html lang="en">
-      <body className="font-pretendard">
-        <div className="flex">
-          <Navbar />
-          <div className="flex flex-col w-full ml-[360px]">
-            <Header />
+      <body>
+        <Layout>
+          <div>
             <main>{children}</main>
           </div>
-        </div>
-      </body>
-    </html>
-  );
+        </Layout>
+      </body>  
+      </html>
+    );
 }
